[10.0][MIG] mrp_brewing: Migration to 10.0 - Followed OCA's guidelines: https://github.com/OCA/maintainer-tools/wiki/Migration-to-version-10.0 - NOT tested on 10.0
