+++ mrp_brewing/README.rst
@@ -1,0 +1,139 @@
+===========
+MRP Brewing
+===========
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Faddons-lightgray.png?logo=github
+    :target: https://github.com/OCA/addons/tree/10.0/mrp_brewing
+    :alt: OCA/addons
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/addons-10-0/addons-10-0-mrp_brewing
+    :alt: Translate me on Weblate
+
+|badge1| |badge2| |badge3| |badge4| 
+
+This module allows to manage breweries.
+It extends standard mrp modules create recipes for beers in different steps.
+We get the tracability of every product that went into a finished beer.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Usage
+=====
+
+Models
+~~~~~~~
+
+mrp.bom - Bill of Material
+--------------------------
+
+The Bill of Material (BOM) describes the ingredients needed to create a product.
+Here, it will only be a step in the general beer recipe.
+For example it will tell you that you need 1g of yeast and 1hL of *mout* (?)
+to create 1hL of *green* beer.
+
+From the BOM, the user can create Manufacturing Orders.
+
+NB: Bill of Material is *Nomenclature* in French.
+
+mrp.production - Manufacturing Order
+------------------------------------
+
+Created from a BOM template, a Manufacturing Order (MO) is an actual step of production.
+The steps are:
+
+1. New
+    - Instanciated from the BOM
+2. Confirmed - Awaiting for raw material
+    - ``move_lines`` are created for incoming materials
+    - The button *book resources* (?) allows to book the needed resources.
+3. Ready
+    - Once all resources are booked, the MO goes to ready state.
+    - Through the "Produce" button, moves the MO to the production state.
+4. In Production
+    - While producing the step, the brewer moves the ingredients from
+      the *to consume* to the *consumed* columns by allocating a lot number to each
+      product.
+5. Done
+6. Cancelled
+
+
+brew.order - Brew Order
+-----------------------
+
+brew.declaration - Brew Declaration
+-----------------------------------
+
+Master Manufacturing Order
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Many steps are needed to produce a beer.
+We want to link all the steps for a single beer together in order to:
+
+* symplify the creation of the MO:
+  from a *Master MO* (eg. Green [#]_ Pale Ale), all linked steps (ie. MO) are created.
+* Link all these MO to a single lot number.
+  The lot number is created for the Master Manufacturing Order and then propa
+
+-> I still don't really get how lot number is linked to MO...
+
+.. [#] The *Green* beer is an unmatured beer.
+
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/addons/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/addons/issues/new?body=module:%20mrp_brewing%0Aversion:%2010.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Coop IT Easy SCRLfs
+
+Contributors
+~~~~~~~~~~~~
+
+* Open Architects Consulting SPRL
+* `Coop IT Easy SCRLfs <https://coopiteasy.be>`_:
+
+  * Houssine Bakkali <houssine@coopiteasy.be>
+  * Vincent Van Rossem <vincent@coopiteasy.be>
+
+
+Maintainers
+~~~~~~~~~~~
+
+This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+This module is part of the `OCA/addons <https://github.com/OCA/addons/tree/10.0/mrp_brewing>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
