+++ mrp_brewing/__manifest__.py
@@ -1,17 +1,15 @@
-# -*- coding: utf-8 -*-
 # © 2016-2018 Open Architects Consulting SPRL.
 # © 2018 Coop IT Easy SCRLfs. (<http://www.coopiteasy.be>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
     "name": "MRP Brewing",
-    "description": "This module allows to handle product transformation",
+    "summary": "This module allows to handle product transformation",
     "category": "Stock",
-    "version": "9.0.1.0.0",
+    "version": "10.0.1.0.0",
     "author": "Coop IT Easy SCRLfs",
-    "website": "www.coopiteasy.be",
+    "website": "https://coopiteasy.be",
     "license": "AGPL-3",
     "depends": [
-        "base",
         "mrp_byproduct",
         "stock",
         "sale",
@@ -30,14 +28,23 @@
         "report/report_layout.xml",
         "wizard/stock_reports_view.xml",
         "wizard/recompute_qty_afte_move_view.xml",
-        "views/product_view.xml",
-        "views/mrp_view.xml",
-        "views/brew_view.xml",
-        "views/stock_view.xml",
-        "views/sale_view.xml",
+        "views/product_template_views.xml",
+        "views/purchase_views.xml",
+        "views/product_pricelist_views.xml",
+        "views/mrp_production_views.xml",
+        "views/brew_declaration_views.xml",
+        "views/brew_order_views.xml",
+        "views/stock_warehouse_views.xml",
+        "views/stock_quant_views.xml",
+        "views/stock_production_lot_views.xml",
+        "views/stock_pack_operation_views.xml",
+        "views/stock_picking_views.xml",
+        "views/stock_move_views.xml",
+        "views/sale_views.xml",
         "views/sale_order_report.xml",
-        "views/partner_view.xml",
-        "views/company_view.xml",
+        "views/res_partner_view.xml",
+        "views/res_company_view.xml",
     ],
     "application": True,
+    "installable": True,
 }
