+++ mrp_brewing/data/cron.xml
@@ -1,27 +1,23 @@
 <odoo>
-    <data noupdate="0">
-        <record id="ir_cron_batch_fix_stock_production_lot_quantity" model="ir.cron">
-            <field name="name">Stock Production Lot - Batch Fix Quantity Available</field>
-            <field name="interval_number">12</field>
-            <field name="interval_type">months</field>
-            <field name="numbercall">-1</field>
-            <field name="doall" eval="False" />
-            <field name="model">stock.production.lot</field>
-            <field name="function">_batch_compute_qty_available</field>
-            <field name="args">()</field>
-        </record>
-    </data>
+    <record id="ir_cron_batch_fix_stock_production_lot_quantity" model="ir.cron">
+        <field name="name">Stock Production Lot - Batch Fix Quantity Available</field>
+        <field name="interval_number">12</field>
+        <field name="interval_type">months</field>
+        <field name="numbercall">-1</field>
+        <field name="doall" eval="False" />
+        <field name="model">stock.production.lot</field>
+        <field name="function">_batch_compute_qty_available</field>
+        <field name="args">()</field>
+    </record>
 
-    <data noupdate="0">
-        <record id="ir_cron_batch_partner_compute_sales_statistics" model="ir.cron">
-            <field name="name">MRP Brewing - Batch Compute Sale Statistics</field>
-            <field name="interval_number">1</field>
-            <field name="interval_type">days</field>
-            <field name="numbercall">-1</field>
-            <field name="doall" eval="False" />
-            <field name="model">res.partner</field>
-            <field name="function">_batch_compute_sales_statistics</field>
-            <field name="args">()</field>
-        </record>
-    </data>
+    <record id="ir_cron_batch_partner_compute_sales_statistics" model="ir.cron">
+        <field name="name">MRP Brewing - Batch Compute Sale Statistics</field>
+        <field name="interval_number">1</field>
+        <field name="interval_type">days</field>
+        <field name="numbercall">-1</field>
+        <field name="doall" eval="False" />
+        <field name="model">res.partner</field>
+        <field name="function">_batch_compute_sales_statistics</field>
+        <field name="args">()</field>
+    </record>
 </odoo>
