+++ mrp_brewing/models/__init__.py
@@ -1,7 +1,11 @@
 from . import product
-from . import stock
-from . import mrp
-from . import brew
+from . import stock_move
+from . import stock_picking
+from . import stock_production_lot
+from . import stock_warehouse
+from . import mrp_production
+from . import brew_declaration
+from . import brew_order
 from . import sale
-from . import partner
-from . import company
+from . import res_partner
+from . import res_company
