+++ mrp_brewing/models/stock_move.py
@@ -1,8 +1,7 @@
-# -*- coding: utf-8 -*-
 # Part of Open Architechts Consulting sprl. See LICENSE file for full copyright and licensing details. # noqa
 # Copyright 2019 Coop IT Easy SCRLfs
 
-from openerp import api, fields, models
+from odoo import api, fields, models
 
 
 class StockMove(models.Model):
@@ -10,7 +9,7 @@ class StockMove(models.Model):
 
     @api.model
     @api.depends("state")
-    def get_on_hand(self):
+    def _compute_on_hand(self):
         for move in self:
             if move.state == "done":
                 move.quantity_after_move = move.product_id.qty_available
@@ -22,7 +21,10 @@ class StockMove(models.Model):
         readonly=True,
     )
     quantity_after_move = fields.Integer(
-        string="Quantity", compute="get_on_hand", store=True, readonly=True
+        string="Quantity",
+        compute="_compute_on_hand",
+        store=True,
+        readonly=True,
     )
     brew_number = fields.Integer(string="Brew number", readonly=True)
     is_internal = fields.Boolean(
@@ -84,70 +86,3 @@ class StockMove(models.Model):
                 result["fields"].get("is_internal", False)
             )
         return result
-
-
-class StockWarehouseOrderpoint(models.Model):
-    _inherit = "stock.warehouse.orderpoint"
-
-    qty_available = fields.Float(
-        related="product_id.qty_available",
-        string="Quantity On Hand",
-        readonly=True,
-    )
-
-
-class StockProductionLot(models.Model):
-    _inherit = "stock.production.lot"
-    _order = "has_stock desc, create_date asc"
-
-    qty_available = fields.Float(
-        string="Quantity available",
-        compute="_compute_qty_available",
-        store=True,
-    )
-
-    has_stock = fields.Boolean(
-        string="Has Stock", compute="_compute_qty_available", store=True
-    )
-
-    @api.multi
-    @api.depends("quant_ids.reservation_id")
-    def _compute_qty_available(self):
-        for lot in self:
-            quants = lot.quant_ids.filtered(
-                lambda r: r.location_id.usage == "internal"
-                and not r.reservation_id
-            )  # noqa
-            qty = sum(quants.mapped("qty"))
-            lot.qty_available = qty
-            lot.has_stock = qty > 0
-
-    @api.model
-    def _batch_compute_qty_available(self):
-        lots = self.env["stock.production.lot"].search([])
-        lots._compute_qty_available()
-
-
-class StockPicking(models.Model):
-    _inherit = "stock.picking"
-
-    date = fields.Datetime(
-        "Creation Date",
-        help="Creation Date, usually the time of the order",
-        select=True,
-        states={
-            "done": [("readonly", False)],
-            "cancel": [("readonly", False)],
-        },
-        track_visibility="onchange",
-    )
-    date_done = fields.Datetime(
-        readonly=False, states={"done": [("readonly", True)]}
-    )
-
-    @api.multi
-    def do_transfer(self):
-        super(StockPicking, self).do_transfer()
-        for picking in self:
-            for move in picking.move_lines:
-                move.date = picking.date_done
