+++ mrp_brewing/readme/CONTRIBUTORS.rst
@@ -1,0 +1,5 @@
+* Open Architects Consulting SPRL
+* `Coop IT Easy SCRLfs <https://coopiteasy.be>`_:
+
+  * Houssine Bakkali <houssine@coopiteasy.be>
+  * Vincent Van Rossem <vincent@coopiteasy.be>
