+++ mrp_brewing/readme/DESCRIPTION.rst
@@ -1,0 +1,3 @@
+This module allows to manage breweries.
+It extends standard mrp modules create recipes for beers in different steps.
+We get the tracability of every product that went into a finished beer.
