+++ mrp_brewing/report/report_finished_products.py
@@ -1,7 +1,6 @@
-# -*- coding: utf-8 -*-
-
 import time
-from openerp import api, models
+
+from odoo import api, models
 
 
 class ReportFinishedProducts(models.AbstractModel):
