--- mrp_brewing/views/company_view.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" ?>
-<odoo>
-    <data>
-        <record id="view_company_inherit_form" model="ir.ui.view">
-            <field name="name">res.company.form.inherit</field>
-            <field name="model">res.company</field>
-            <field name="inherit_id" ref="base.view_company_form"/>
-            <field name="arch" type="xml">
-                <field name="company_registry" position="after">
-                    <field name="warehousekeeper_no"/>
-                </field>
-            </field>
-        </record>
-    </data>
-</odoo>
