--- mrp_brewing/views/mrp_view.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-	<data>
-		<!-- <record id="view_mrp_bom_line_filter_lot" model="ir.ui.view">
-            <field name="name">mrp.bom.line.select</field>
-            <field name="model">mrp.bom.line</field>
-            <field name="inherit_id" ref="mrp.view_mrp_bom_line_filter_lot"/>
-            <field name="arch" type="xml">
-            	<field name="product_tmpl_id" position="before">
-            		<field name="lot_number"/>
-            	</field>
-            </field>
-        </record> -->
-        
-        <record id="mrp_production_brew_form_view" model="ir.ui.view">
-            <field name="name">mrp.production.brew.form</field>
-            <field name="model">mrp.production</field>
-            <field name="inherit_id" ref="mrp.mrp_production_form_view"/>
-            <field name="arch" type="xml">
-            	<field name="origin" position="before">
-            		<field name="brew_orders" invisible="True"/>
-            		<field name="master_mo_id"/>
-            		<field name="brew_number"/>
-            	</field>
-            </field>
-        </record>
-	</data>
-</odoo>
