+++ mrp_brewing/views/product_template_views.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="intern_code_form_view" model="ir.ui.view">
+        <field name="name">product.template.product.form</field>
+        <field name="model">product.template</field>
+        <field name="inherit_id" ref="product.product_template_only_form_view" />
+        <field name="arch" type="xml">
+            <field name="default_code" position="replace">
+                <field
+                    name="default_code"
+                    attrs="{'invisible': [('product_variant_count', '>', 1)],'required':[('is_brewable','=',True)]}"
+                />
+            </field>
+        </field>
+    </record>
+</odoo>
