--- mrp_brewing/views/product_view.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-	<data>
-        <record id="intern_code_form_view" model="ir.ui.view">
-            <field name="name">product.template.product.form</field>
-            <field name="model">product.template</field>
-            <field name="inherit_id" ref="product.product_template_only_form_view"/>
-            <field name="arch" type="xml">
-            	<field name="default_code" position="replace">
-            		<field name="default_code" attrs="{'invisible': [('product_variant_count', '>', 1)],'required':[('is_brewable','=',True)]}"/>
-            	</field>
-            </field>
-        </record>
-        
-        <record id="raw_material_purchase_product_template_form" model="ir.ui.view">
-            <field name="name">product.template.raw</field>
-            <field name="model">product.template</field>
-            <field name="inherit_id" ref="purchase.view_product_supplier_inherit"/>
-            <field name="arch" type="xml">
-                <xpath expr="//label[@for='purchase_ok']/.." position="after">
-                    <div>
-                        <field name="raw_material"/>
-                        <label for="raw_material"/>
-                    </div>
-                    <div>
-                        <field name="finished_product"/>
-                        <label for="finished_product"/>
-                    </div>
-                    <div>
-                        <field name="is_brewable"/>
-                        <label for="is_brewable"/>
-                    </div>
-                    <div>
-                        <field name="is_crate"/>
-                        <label for="is_crate"/>
-                    </div>
-                </xpath>
-                <field name="type" position="after">
-                	<field name="brew_product_sequence" attrs="{'invisible':[('is_brewable','=',False)],'required':[('is_brewable','=',True)]}"/>
-                </field>
-            </field>
-        </record>
-        
-        <record id="product_pricelist_view" model="ir.ui.view">
-            <field name="name">product.pricelist.form</field>
-            <field name="model">product.pricelist</field>
-            <field name="inherit_id" ref="product.product_pricelist_view"/>
-            <field name="arch" type="xml">
-            	<field name="company_id" position="after">
-            		<field name="particular_conditions"/>
-            	</field>
-        	</field>
-        </record>
-    </data>
-</odoo>
