+++ mrp_brewing/views/sale_order_report.xml
@@ -1,15 +1,15 @@
-<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-	<data>
-		<template id="report_saleorder_brewing" inherit_id="sale.report_saleorder_document">
-            <xpath expr="//tbody[@class='sale_tbody']/../thead/tr/th[1]" position="after">
-				<th class="text-right">Production lot</th>
-            </xpath>
-			<xpath expr="//span[@t-field='l.name']" position="after">
-				<td class="text-right">
-		            <span t-esc="', '.join(map(lambda x: (x.name or x.ref), l.product_lot_ids))"/>
-		        </td>
-	        </xpath>
-	  	</template>
-	</data>
+    <template id="report_saleorder_brewing" inherit_id="sale.report_saleorder_document">
+        <xpath expr="//tbody[@class='sale_tbody']/../thead/tr/th[1]" position="after">
+            <th class="text-right">Production lot</th>
+        </xpath>
+        <xpath expr="//span[@t-field='l.name']" position="after">
+            <td class="text-right">
+                <span
+                    t-esc="', '.join(map(lambda x: (x.name or x.ref), l.product_lot_ids))"
+                />
+            </td>
+        </xpath>
+    </template>
 </odoo>
