--- mrp_brewing/views/sale_view.xml
@@ -1,83 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-	<data>
-		
-		<record id="sale.action_quotations" model="ir.actions.act_window">
-            <field name="name">Quotations</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">sale.order</field>
-            <field name="view_type">form</field>
-            <field name="view_id" ref="sale.view_quotation_tree"/>
-            <field name="view_mode">tree,kanban,form,calendar,pivot,graph</field>
-            <field name="context">{'search_default_draft': True}</field>
-            <field name="domain">[('state','in',('draft','sent','cancel'))]</field>
-            <field name="search_view_id" ref="sale.view_sales_order_filter"/>
-            <field name="help" type="html">
-              <p class="oe_view_nocontent_create">
-                Create a Quotation, the first step of a new sale.
-              </p><p>
-                Your next actions should flow efficiently: confirm the Quotation 
-                to a Sale Order, then create the Invoice and collect the Payment.
-              </p><p>
-                Note that once a Quotation becomes a Sale Order, it will be moved 
-                from the Quotations list to the Sales Order list.
-              </p>
-            </field>
-        </record>
-        
-        <record id="view_order_tree" model="ir.ui.view">
-            <field name="name">sale.order.tree</field>
-            <field name="model">sale.order</field>
-            <field name="inherit_id" ref="sale.view_order_tree"/>
-            <field name="arch" type="xml">
-              <xpath expr="//field[@name='amount_total']" position="attributes">
-                <attribute name="invisible">True</attribute>
-              </xpath>
-              <field name="amount_total" position='after'>
-                <field name="amount_untaxed"/>
-              </field>
-            	<field name="invoice_status" position="after">
-            		<field name="delivery_done_date"/>
-            		<field name="delivery_status"/>
-            	</field>
-            </field>
-        </record>
-		
-		<record id="view_order_form" model="ir.ui.view">
-            <field name="name">sale.order.form</field>
-            <field name="model">sale.order</field>
-            <field name="inherit_id" ref="sale.view_order_form"/>
-            <field name="arch" type="xml">
-            	<xpath expr="//field[@name='order_line']/tree/field[@name='name']" position="after">
-            		<field name="product_lot_ids" widget="many2many_tags"/>
-            	</xpath>
-            </field>
-        </record>
-        
-        <record id="view_order_line_tree_brew" model="ir.ui.view">
-            <field name="name">sale.order.line.tree</field>
-            <field name="model">sale.order.line</field>
-            <field name="inherit_id" ref="sale.view_order_line_tree"/>
-            <field name="arch" type="xml">
-            	<field name="name" position="after">
-            		<field name="product_lot_ids"/>
-            		<field name="effective_date"/>
-            	</field>
-            </field>
-        </record>
-
-    <record id="view_article_month_pivot" model="ir.ui.view">
-     <field name="name">sale.article.month.pivot</field>
-     <field name="model">sale.report</field>
-     <field name="arch" type="xml">
-       <pivot string="Sales Analysis" disable_linking="True">
-         <field name="product_id" type="row"/>
-         <field name="date" interval="month" type="col"/>
-         <field name="qty_invoiced" type="measure"/>
-         <field name="price_total" type="measure"/>
-       </pivot>
-     </field>
-    </record>
-
-	</data>
-</odoo>
