+++ mrp_brewing/views/stock_move_views.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="view_move_picking_tree" model="ir.ui.view">
+        <field name="name">stock.move.form</field>
+        <field name="model">stock.move</field>
+        <field name="inherit_id" ref="stock.view_move_picking_tree" />
+        <field name="arch" type="xml">
+            <field name="location_id" position="after">
+                <field name="is_internal" invisible="1" />
+            </field>
+        </field>
+    </record>
+</odoo>
