--- mrp_brewing/views/stock_view.xml
@@ -1,100 +1,0 @@
-<odoo>
-	<data>
-		<record id="view_warehouse_orderpoint_tree" model="ir.ui.view">
-            <field name="name">stock.warehouse.orderpoint.tree</field>
-            <field name="model">stock.warehouse.orderpoint</field>
-            <field name="inherit_id" ref="stock.view_warehouse_orderpoint_tree"/>
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='name']/.." position="replace">
-	                <tree string="Reordering Rules" colors="red:qty_available &lt; product_min_qty">
-	                    <field name="warehouse_id" groups="stock.group_locations"/>
-	                    <field name="location_id" groups="stock.group_locations"/>
-	                    <field name="product_id"/>
-	                    <field name="product_uom" options="{'no_open': True, 'no_create': True}" groups="product.group_uom"/>
-	                	<field name="qty_available"/>
-	                    <field name="product_min_qty"/>
-	                    <field name="product_max_qty"/>
-	                </tree>
-                </xpath>
-            </field>
-        </record>
-        
-        <record id="quant_search_view" model="ir.ui.view">
-            <field name="name">stock.quant.search</field>
-            <field name="model">stock.quant</field>
-            <field name="inherit_id" ref="stock.quant_search_view"/>
-            <field name="arch" type="xml">
-            	<xpath expr="//filter[@groups='stock.group_production_lot']" position="replace">
-            		<filter name="lot" string='Lot' context="{'group_by': 'lot_id'}" groups="stock.group_production_lot"/>
-             	</xpath>
-             	<xpath expr="//filter[@groups='stock.group_production_lot']/.." position="attributes">
-                    <attribute name="expand">1</attribute>
-                </xpath>
-            </field>
-        </record>
-
-        <record model="ir.actions.act_window" id="sale_product_open_quants">
-            <field name="name">Stock On Hand</field>
-            <field name="context">{'search_default_internal_loc': 1, 'search_default_locationgroup':1, 'search_default_productgroup':1,'search_default_lot':1}</field>
-            <field name="domain">[('product_id.finished_product', '=', True)]</field>
-            <field name="res_model">stock.quant</field>
-        </record>
-        
-        <menuitem action="sale_product_open_quants" id="menu_sale_product_open_quants_action" parent="stock.menu_stock_inventory_control" sequence="10"/>
-        
-        <record id="view_production_lot_form" model="ir.ui.view">
-            <field name="name">stock.production.lot.form</field>
-            <field name="model">stock.production.lot</field>
-            <field name="inherit_id" ref="stock.view_production_lot_form"/>
-            <field name="arch" type="xml">
-            	<field name="ref" position="after">
-            		<field name="qty_available"/>
-            	</field>
-            </field>
-        </record>
-        
-        <record id="view_production_lot_tree" model="ir.ui.view">
-            <field name="name">stock.production.lot.tree</field>
-            <field name="model">stock.production.lot</field>
-            <field name="inherit_id" ref="stock.view_production_lot_tree"/>
-            <field name="arch" type="xml">
-            	<field name="product_id" position="after">
-            		<field name="qty_available"/>
-            	</field>
-            </field>
-        </record>
-        
-        <record id="view_pack_operation_lot_form" model="ir.ui.view">
-            <field name="name">stock.pack.operation.lots.form</field>
-            <field name="model">stock.pack.operation</field>
-            <field name="inherit_id" ref="stock.view_pack_operation_lot_form"/>
-            <field name="arch" type="xml">
-            	<field name="lot_id" position="attributes">
-            		<attribute name="options">{'no_create': True}</attribute>
-            	</field>
-            </field>
-        </record>
-        
-        <record id="view_picking_form" model="ir.ui.view">
-            <field name="name">stock.picking.form</field>
-            <field name="model">stock.picking</field>
-            <field name="inherit_id" ref="stock.view_picking_form"/>
-            <field name="arch" type="xml">
-		        <field name="min_date" position="before">
-		        	<field name="date_done"/>
-		        </field>
-            </field>
-        </record>
-
-        <record id="view_move_picking_tree" model="ir.ui.view">
-            <field name="name">stock.move.form</field>
-            <field name="model">stock.move</field>
-            <field name="inherit_id" ref="stock.view_move_picking_tree"/>
-            <field name="arch" type="xml">
-                <field name="location_id" position="after">
-                    <field name="is_internal" invisible="1"/>
-                </field>
-            </field>
-        </record>
-	</data>
-</odoo>
