+++ mrp_brewing/views/stock_warehouse_views.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="view_warehouse_orderpoint_tree" model="ir.ui.view">
+        <field name="name">stock.warehouse.orderpoint.tree</field>
+        <field name="model">stock.warehouse.orderpoint</field>
+        <field name="inherit_id" ref="stock.view_warehouse_orderpoint_tree" />
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='name']/.." position="replace">
+                <tree
+                    string="Reordering Rules"
+                    colors="red:qty_available &lt; product_min_qty"
+                >
+                    <field name="warehouse_id" groups="stock.group_locations" />
+                    <field name="location_id" groups="stock.group_locations" />
+                    <field name="product_id" />
+                    <field
+                        name="product_uom"
+                        options="{'no_open': True, 'no_create': True}"
+                        groups="product.group_uom"
+                    />
+                    <field name="qty_available" />
+                    <field name="product_min_qty" />
+                    <field name="product_max_qty" />
+                </tree>
+            </xpath>
+        </field>
+    </record>
+</odoo>
