+++ mrp_brewing/wizard/mrp_produce_wizard.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-from openerp import api, fields, models
+from odoo import api, fields, models
 
 
 class MRPProductProduce(models.TransientModel):
