+++ setup/account_invoice_check_bba_com/odoo/addons/account_invoice_check_bba_com
@@ -1,0 +1,1 @@
+../../../../account_invoice_check_bba_com
