+++ setup/mrp_brewing/odoo/addons/mrp_brewing
@@ -1,0 +1,1 @@
+../../../../mrp_brewing
